+++ ResxFormatter/ISettings.cs
@@ -1,9 +1,16 @@
 ﻿namespace ResxFormatter
 {
+    public enum ReloadMode
+    {
+        Off,
+        AfterModification,
+        Always
+    }
+
     public interface ISettings
     {
         bool FixResxWriter { get; }
-        bool ReloadFileAutomatically { get; }
+        ReloadMode ReloadFile { get; }
         bool RemoveDocumentationComment { get; }
         bool SortEntries { get; }
     }