+++ ResxFormatter/ResxFormatterPackage.cs
@@ -89,7 +89,9 @@
             {
                 Log.WriteLine("Save event for xml document received.");
                 var formatter = new ResxFormatter(settings, Log);
-                if (formatter.Run(document.FullName) && settings.ReloadFileAutomatically)
+                if ((formatter.Run(document.FullName) && settings.ReloadFile == ReloadMode.Off)
+                    || settings.ReloadFile == ReloadMode.Always)
+
                 {
                     Log.WriteLine("Reloading file.");
                     document.Close(vsSaveChanges.vsSaveChangesNo);